requirements.txt
opencv-python>=4.5.1.48
numpy>=1.19.5
torch>=1.7.1+cu110
Pillow>=8.1.0
tqdm>=4.56.0
scipy>=1.3.3
scikit-image>=0.17.2
munkres>=1.1.4
torchvision>=0.8.2+cu110
typing>=3.7.4.3